+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0 2020-09-21
+
+(not backwards compatible)
+
+- make `pmap` return a generator, so that the entire result set does
+  not have to be in memory at the same time
+
 ## 0.1.1 2020-09-20
 
 - Remove deprecated `loop` argument from `asyncio.wait` (fixes warning)